requirements.txt
altair==5.3.0
joblib==1.4.2
matplotlib==3.9.0
plotly==5.22.0
polars==0.20.31
pygwalker==0.4.8.9
PyYAML==6.0.1
st-annotated-text==4.0.1
statsmodels==0.14.1
streamlit==1.35.0